--- a/src/libraries/System.Net.Http/src/System/Net/Http/HttpMethod.cs
+++ b/src/libraries/System.Net.Http/src/System/Net/Http/HttpMethod.cs
@@ -176,9 +176,12 @@ namespace System.Net.Http
         /// </summary>
         internal static HttpMethod Normalize(HttpMethod method)
         {
+            // _http3EncodedBytes is only set for the singleton instances, so if it's not null,
+            // we can avoid the dictionary lookup.  Otherwise, look up the method instance in the
+            // dictionary and return the normalized instance if it's found.
             Debug.Assert(method != null);
-            return s_knownMethods.TryGetValue(method, out HttpMethod? normalized) ?
-                normalized :
+            return
+                method._http3EncodedBytes is null && s_knownMethods.TryGetValue(method, out HttpMethod? normalized) ? normalized :
                 method;
         }