… src/libraries/System.Net.Http/src/System/Net/Http/SocketsHttpHandler/HttpConnectionBase.cs
… src/libraries/System.Net.Http/src/System/Threading/AsyncMutex.cs
… src/libraries/System.Net.Http/tests/UnitTests/Headers/KnownHeadersTest.cs