@stephentoub 2020-05-01
> This seems worrisome from a security standpoint. Can we consider at a minimum using a different ArrayPool than the shared one to mitigate the possibility that code elsewhere (mis-)using ArrayPool might end up leaking sensitive data?

This is one of many places `ArrayPool<byte>.Shared.Return` is used without clearing, especially since not clearing is the default:
https://source.dot.net/#System.Private.CoreLib/ArrayPool.cs,34fc80a1ec3f75d7,references
That includes other places in this class:
https://source.dot.net/#System.Net.Http/ArrayBuffer.cs,129
In other words, this particular call site was an anomaly.

If we want to revisit such clearing decisions, we can, but that should be a separate issue / PR.  This change is really just bringing this particular use into consistency.