@davidfowl 2020-05-01
> Also, to my knowledge, most of pipelines and more generally Kestrel isn't clearing for its pooling, its slab memory blocks, etc. @davidfowl can correct me if I'm wrong.

That's right. It's more than even pipelines, most of ASP.NET isn't clearing the memory on return in general. That goes for pipelines and other uses of the array pool.