@JamesNK 2020-05-01
Nice! That's a great improvement.

Would it be useful to you to have benchmarks from different gRPC frameworks? e.g. golang and/or Java. I haven't looked closely at gRPC client performance until now, and I'm not sure whether Grpc.Core client is a good target to aim towards.

Grpc.Core performance might be very good compared to the other gRPC clients because it is closely tied to its internal HTTP/2 client. Meanwhile the Java gRPC client is built on a general purpose HTTP/2 client: [OkHttp](https://square.github.io/okhttp/).

On the other hand Grpc.Core performance might be bad compared to other gRPC clients because it has the overhead of native execution and then .NET wrapper on top.