@aaronla-ms 2020-05-01
Seconding @davidni 's comment re memory clearing, but concurring with @stephentoub -- this is a high level security design decision that should be made at a project level, not individual PRs. It'll be most effective when coordinated consistently across a product.

Re prior art, I don't know of any pipeline that clears its buffers, but I do know of security defects whose impact was worsened by not clearing buffers. The most recent that comes to mind was one where a particular DDoS protection gateway leaked traffic between connections. Clearing sensitive buffers before reuse is one way to preemptively mitigate such security failures.