@stephentoub 2020-05-01
> Would it be useful to you to have benchmarks from different gRPC frameworks? e.g. golang and/or Java.

Yeah, @jamesNK, that'd be great.