@stephentoub 2020-05-05
I ran HttpStress on this for about a half-hour locally, and while there were a 11 failed requests out of about a million, that's fewer failures than when I ran it for a similar period of time without my changes.