@geoffkizer 2020-06-02
> A "header serialization lock" was being employed and was a significant source of contention. This lock protected a shared header buffer. Rather than maintaining such a buffer on the Http2Connection, we instead just rent a pool buffer to write the headers into, which means we no longer need to serialize access to the shared buffer. The lock has been deleted.

Note that part of the reason this existed in the first place was to serialize state changes within the header encoder, so that we could potentially add dynamic header encoding in the future.

I strongly doubt we will add dynamic header encoding here in the future, so it's probably a non-issue.